Remove voltage and movement info from PlotDataPage
As now all the latest values can be seen in the SelectDevicePage,
it's unneccesary to show these in here too. The page would get a
lot of information if we would show all the latest values in here,
and we would need to dynamically update those whenever new readings
comes in so decided to remove the voltage and movement counter from
here.